Extract a base class
