+++ lib/bump/cli/commands/preview.rb
@@ -3,7 +3,7 @@ require 'open-uri'
 module Bump
   class CLI
     module Commands
-      class Preview < Hanami::CLI::Command
+      class Preview < Base
         desc "Create a documentation preview for the given file"
         argument :file, required: true, desc: "Path or URL to your API documentation file. Only OpenApi 2.0 (Swagger) specification is currently supported."
         option :format, default: "yaml", values: %w[yaml json], desc: "Format of the definition"
@@ -25,16 +25,6 @@ module Bump
 
         private
 
-        def headers(options)
-          default_headers
-        end
-
-        def default_headers
-          {
-            "Content-Type" => "application/json"
-          }
-        end
-
         def body(options)
           {
             preview: {
@@ -42,22 +32,6 @@ module Bump
               format: options.fetch(:format)
             }
           }
-        end
-
-        def display_invalid_definition(body)
-          puts "Definition is not valid:"
-          body["errors"]["raw_definition"].each do |error|
-            puts "> #{error}"
-          end
-          abort
-        end
-
-        def display_generic_error(body)
-          abort body["message"]
-        end
-
-        def display_http_error(error)
-          abort "#{error.class}: #{error.message}"
         end
       end
     end
